refactor: refactor settings
